Added `NPM Audit Report`
This action audits the `package.json` of a project and comments on the associated PR to highlight any nasty dependency issues that may have arisen recently. It is intended to provide visibility on the open audit issues rather than mandate an update/fix, leaving enforcing these issues to the PR approver/raiser. It runs on a `pull request` trigger and executes `npm audit` on the specified `package.json`.